feature(lib): move Input transform to retrieval from pyproject.toml
Input transform is mostly used to keep the internal Input value in the
same format as pyproject.toml, so that it's then easier to write it in
the file.
However, it's actually simpler to keep the internal Input value to
whatever is set and only convert it to the format it has in
pyproject.toml in the generator that writes the file.
However, this implies that the value read from pyproject.toml is
transformed to the internal Input value format upon retrieval from
pyproject.toml